Learning Git glow blow snow flow
Oh no we need to keep additional lines to less than 75 chars. But
whatever happened to the 72 char limit?
The test is modified to allow a negative balance on the test. This
could lead to more features on accounts where if the customer has
overdraft protection then an additional nasty fee it also applied to the
account. Or another option would be to have the feature which prevents an
overdraft from happening on the customer's account.

Change-Id: Ic75cb3c43ab25ce7c8eda578268c214d29a0034e